Finalized versions of Test3 and Test4
Added Test3 and Test4. Test3 is the baseline test java file. It simply creates an array of integers 1-100. It then randomizes the order of the array. Finally, it uses a tree map to re-sort the array. It can be modified for any amount of integers. Test4 is identical but adds print statements for all of the relevant debugging lines. Finally, there are logs to indicate all testing done (documented on spreadsheet).